# Poor batching performance reproducer