… examples/bench/go.sum