This is a reproducer for https://github.com/twmb/franz-go/issues/1063

With franz-go single topic single partition:

```
go run . -topic foo -compression=snappy      
99.83 MiB/s; 1046.82k records/s
104.14 MiB/s; 1091.97k records/s
99.89 MiB/s; 1047.47k records/s
100.27 MiB/s; 1051.36k records/s
102.20 MiB/s; 1071.63k records/s
102.28 MiB/s; 1072.50k records/s
103.64 MiB/s; 1086.77k records/s
99.16 MiB/s; 1039.79k records/s
100.10 MiB/s; 1049.67k records/s
99.75 MiB/s; 1045.91k records/s
101.58 MiB/s; 1065.16k records/s
```

With segmentio:

```
go run . -topic foo -compression=snappy -fast
198.36 MiB/s; 2080.00k records/s
214.58 MiB/s; 2250.00k records/s
216.31 MiB/s; 2268.20k records/s
206.17 MiB/s; 2161.80k records/s
207.90 MiB/s; 2180.00k records/s
212.67 MiB/s; 2230.00k records/s
220.30 MiB/s; 2310.00k records/s
206.95 MiB/s; 2170.00k records/s
210.76 MiB/s; 2210.00k records/s
214.58 MiB/s; 2250.00k records/s
206.95 MiB/s; 2170.00k records/s
```

I did not try particularly hard to optimize this, .e.g `make([]kafka.Message, len(batch))` could be optimized away but it is still twice as fast...