@twmb 2025-07-30
> normally Redis/Postgres/Clickhouse etc use a pool of connections so users don't have to worry about the details.

This is opposite of what is prescribed per Kafka clients. Kafka clients generally only use one, while opening a second (or third or fourth) depending on what you're doing. Kafka clients are super heavy weight internally and are _meant_ to do as much batching as possible and send as much as possible down a single connection. I did evaluate actually having a pool of connections in the past, but the thinking at the time was actually to _reduce_ the number of connections open to a broker: https://github.com/twmb/franz-go/issues/121#issuecomment-1055817531

Using multiple connections to _increase_ throughput would pretty quickly run into ordering semantics problems and would likely break idempotent producers.

> does not allow to check if kgo.Client is busy / can accept more messages

One thing you could do is opt into `ManualFlushing`, and then call `TryProduce` for everything. One `TryProduce` returns `ErrMaxBuffered`, you `Flush` and then switch to your second client. You only resume on the first client once Flush returns (since flush hangs until everything is flushed). You may need to work more clients into the rotation depending on how things play out; you could also cancel flushing by quitting the flush context. I'm not saying this is a good idea fwiw, just something to that may be worth trying.

In the mean time, I may have a coworker help me setting up a lot of AI stuff to see if I can have Claude do most of the rewriting ProduceRecords for me. It's a large and a bit tricky change and I'm not too keen to spend hours and hours of my spare time manually doing this for ............... a relatively niche amount of throughput 😅