@twmb 2025-10-07
So, coming back to this: franz-go, for me, has the same performance as segment _without_ any changes. I've added to the `bench` utility to imitate the code you added to segment. If I create 10k recs _before_ calling `Produce` and then produce them all quickly, and if I do that in 5 goroutines (similar to your max-write-threads), I get the same performance.

That said, I also worked on ProduceSyncBatch which has a 50% performance increase over the above.

Results below, running against a local 3 node 4,0 KRaft cluster

```
[18:42:25]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ rpk topic create foo -p 3
TOPIC  STATUS
foo    OK

[18:42:37]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ go run . -topic foo -compression=snappy -pgoros 5 -batch-recs 10000 -psync-batch
300.41 MiB/s; 3150.00k records/s
307.08 MiB/s; 3220.00k records/s
301.36 MiB/s; 3160.00k records/s
308.04 MiB/s; 3230.00k records/s
309.94 MiB/s; 3250.00k records/s
313.76 MiB/s; 3290.00k records/s
^Csignal: interrupt

[19:29:54]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ go run . -topic foo -compression=snappy -pgoros 5 -batch-recs 10000  
170.95 MiB/s; 1792.52k records/s
174.24 MiB/s; 1827.01k records/s
174.54 MiB/s; 1830.19k records/s
172.67 MiB/s; 1810.53k records/s
174.02 MiB/s; 1824.77k records/s
173.56 MiB/s; 1819.88k records/s
^Csignal: interrupt

[19:30:05]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ cd compare/segment/

[19:30:07]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench/compare/segment
$ go run . -topic foo -compression=snappy -max-write-threads 5
175.61 MiB/s; 1841.40k records/s
175.61 MiB/s; 1841.40k records/s
175.61 MiB/s; 1841.40k records/s
178.86 MiB/s; 1875.50k records/s
179.67 MiB/s; 1884.03k records/s
178.05 MiB/s; 1866.97k records/s
^Csignal: interrupt
```