@twmb 2025-10-07
I do recreate your original results without any batching, and adding goroutines without batching makes things worse:
```
[19:37:48]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ go run . -topic foo -compression=snappy -pgoros 5
85.82 MiB/s; 899.91k records/s
84.98 MiB/s; 891.07k records/s
85.60 MiB/s; 897.58k records/s
83.81 MiB/s; 878.82k records/s
^Csignal: interrupt

[19:37:54]
twmb@h4x3r:~/src/twmb/franz-go/examples/bench
$ go run . -topic foo -compression=snappy  
100.79 MiB/s; 1056.81k records/s
100.15 MiB/s; 1050.18k records/s
99.19 MiB/s; 1040.05k records/s
100.17 MiB/s; 1050.37k records/s
99.99 MiB/s; 1048.52k records/s
99.16 MiB/s; 1039.80k records/s
99.79 MiB/s; 1046.38k records/s
^Csignal: interrupt
```