@vmihailenco 2025-10-07
Thanks for working on this! Kafka continues to be the bottleneck, so I'm still interested in
improvements, though I've learned to mitigate it through:

1. Creating more partitions
2. Hosting each partition on a separate broker
3. Using an intermediate buffer and calling `Produce` from a goroutine to avoid blocking the main
   loop

I could switch to `ProduceSyncBatch`, but I'd need to re-implement all the rate-limiting logic that
currently exists in the producer, which will take some time.

**Regarding the API design:**

The new API feels like a low-level building block that requires wrapping with a custom Producer, yet
it's still tightly integrated with franz-go's internal mechanics and batching logic. This creates an
awkward middle ground. It might work in practice as an interim solution, but ideally the existing
`Produce` API would be refactored to address these architectural concerns:

**Client-wide buffering:** If I configure `MaxBufferedRecords(10000)`, it should actually buffer
10,000 records when the client can't keep up with load. [Currently](https://github.com/twmb/franz-go/issues/1081), the "buffer" is just a memory limit for franz-go rather than a true buffer as most developers would expect.

**Per-broker batching:** Each broker connection should have its own buffer for accumulating batches
with separate timeout flushes. Currently, franz-go sends thousands of requests even with lingering
enabled — it should be closer to 1-2 requests per second per broker.

**Async-first design:** The synchronous part of `Produce` should be minimal (e.g., just
rate-limiting), with all heavy work happening asynchronously to avoid blocking the main produce
loop.

I've explored the source code to see if I could implement some of these ideas, but the current
architecture is tightly coupled, making it difficult to know where to start untangling things. So
this is mostly wishful thinking on my part—but hopefully useful feedback for future iterations.

Thanks again for your efforts on this!